Add initial support for arbitrary transforms.
One problem is that these can cause the "run" function to run
forever, either by simply not terminating, or by generating an
infinite number of rows (as part of a cycle).

Another problem is that working with Vector[Any] objects is
painful. The API should provide a wrapper object that makes
records easier to work with.